backend/requirements.txt
flask
together
flask-cors
flask-limiter
gunicorn
requests
python-dotenv
supabase
psycopg2-binary
Flask-Bcrypt
PyJWT
user-agents